Small fixes and additions
Added a line about zooming in the about page.

Hard difficulty is either 16x30 or 30x16 depending on the height and width of the window. If the game is started, the grid won't change(e.g. from 16x30 to 30x16) if your screen becomes wide, but only when you are choosing the difficulty